requirements.txt
fpdf2
streamlit
langchain
langchain-community
langchain-google-genai